LayoutTests/imported/w3c/web-platform-tests/css/css-grid/alignment/grid-align-baseline-flex-004-expected.txt


PASS .target > * 1
PASS .target > * 2
FAIL .target > * 3 assert_equals:
<div data-offset-x="110"><span></span></div>
offsetLeft expected 110 but got 130
PASS .target > * 4
PASS .target > * 5
PASS .target > * 6
PASS .target > * 7
PASS .target > * 8
FAIL .target > * 9 assert_equals:
<div data-offset-x="45"><span></span></div>
offsetLeft expected 45 but got 65
PASS .target > * 10
PASS .target > * 11
PASS .target > * 12
PASS .target > * 13
PASS .target > * 14
PASS .target > * 15
PASS .target > * 16
